change estructure module
