+++ lib/ui/modules/cita/r_asistente_recepcion/pages/cita_list_page.dart
@@ -11,27 +11,30 @@ import 'package:admin_clinica_front/ui/global_widget/page/page_base_desktop.dart
 import 'package:admin_clinica_front/ui/global_widget/page/page_base_phone.dart';
 import 'package:admin_clinica_front/ui/modules/cita/bloc/cita_crear_bloc/cita_create_bloc.dart';
 import 'package:admin_clinica_front/ui/modules/cita/bloc/cita_crear_bloc/cita_create_event.dart';
-import 'package:admin_clinica_front/ui/modules/cita/widget/cita_card_doctor.dart';
-import 'package:admin_clinica_front/ui/modules/cita/widget/doctor_carrusel_card.dart';
+import 'package:admin_clinica_front/ui/modules/cita/r_asistente_recepcion/widgets/cita_card_asist_recep.dart';
+import 'package:admin_clinica_front/ui/modules/cita/r_asistente_recepcion/widgets/doctor_carrusel_card.dart';
 import 'package:admin_clinica_front/ui/modules/doctor/bloc/doctor_list_bloc.dart';
 import 'package:admin_clinica_front/ui/view_models/cita_view/cita_view_models.dart';
 import 'package:admin_clinica_front/ui/view_models/doctor_view/doctor_view_models.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import '../../../../core/di/injections.dart';
-import '../../../../data/datasources/remote/doctor_api.dart';
-import '../../../../dominio/repositories/ilocal_repository.dart';
-import '../../../core/router.dart';
-import '../../../global_widget/page/page_mixin_base.dart';
-import '../bloc/cita_bloc.dart';
+import '../../../../../core/di/injections.dart';
+import '../../../../../data/datasources/remote/doctor_api.dart';
+import '../../../../../dominio/repositories/ilocal_repository.dart';
+import '../../../../core/router.dart';
+import '../../../../global_widget/page/page_mixin_base.dart';
+import '../../bloc/cita_bloc.dart';
 
-class CitaListDoctorPage extends StatelessWidget with ResponsiveWidgetMixin {
-  CitaListDoctorPage({super.key});
+class CitaListAsistenteRecepcionPage extends StatelessWidget with ResponsiveWidgetMixin {
+  CitaListAsistenteRecepcionPage({super.key});
   DateTime dateSelected = DateTime.now();
   DoctorsViewModel? doctorSelected;
   @override
   Widget build(BuildContext context) {
-    return whatIs(context);
+    return BlocProvider(
+      create: (context) => CountIsolateCubit(intervalAction: 60),
+      child: whatIs(context),
+    );
   }
 
   @override
@@ -96,29 +99,80 @@ class CitaListDoctorPage extends StatelessWidget with ResponsiveWidgetMixin {
                           return const Center(child: CircularProgressIndicator());
                         },
                         doctorsLoaded: (stt) {
-                          usuarioBloc.setDoctorSelected(stt.doctors[0]);
                           doctorSelected = usuarioBloc.state.doctorIdSelected;
                           if (doctorSelected != null) {
-                            if (doctorSelected!.isActive) {
-                              context.read<CitaBloc>().add(
-                                    CitaEvent.getCitas(
-                                      CitaRequestViewModel(
-                                        doctorId: doctorSelected!.id,
-                                        ubicacionesId: usuarioBloc.state.usuario?.ubicaciones ?? [],
-                                        fechaHoraCita: DateTime.now(),
+                            if (stt.doctors.any(
+                              (element) => doctorSelected!.id == element.id,
+                            )) {
+                              doctorSelected = stt.doctors.firstWhere((element) => doctorSelected!.id == element.id);
+                              if (doctorSelected!.isActive) {
+                                context.read<CitaBloc>().add(
+                                      CitaEvent.getCitas(
+                                        CitaRequestViewModel(
+                                          doctorId: doctorSelected!.id,
+                                          ubicacionesId: usuarioBloc.state.usuario?.ubicaciones ?? [],
+                                          fechaHoraCita: DateTime.now(),
+                                        ),
                                       ),
-                                    ),
-                                  );
+                                    );
+                              } else {
+                                context.read<CitaBloc>().add(CitaEvent.invalidCita("Este doctor(a) está inactivo"));
+                              }
                             } else {
-                              context.read<CitaBloc>().add(CitaEvent.invalidCita("Este doctor(a) está inactivo"));
+                              if (stt.doctors.isNotEmpty) usuarioBloc.setDoctorSelected(stt.doctors[0]);
                             }
                           }
 
                           doctorSelected = usuarioBloc.state.doctorIdSelected;
-                          return DoctorCarousel(
-                            doctorIdInitialSelected: doctorSelected?.id,
-                            doctors: stt.doctors,
-                            onChanged: (doctor) {},
+                          return BlocListener<CountIsolateCubit, int>(
+                            listener: (context, state) {
+                              if (state % context.read<CountIsolateCubit>().intervalAction == 0) {
+                                timerStart(timerCubit, () {
+                                  if (doctorSelected != null) {
+                                    if (stt.doctors.any(
+                                      (element) => doctorSelected!.id == element.id,
+                                    )) {
+                                      doctorSelected = stt.doctors.firstWhere((element) => doctorSelected!.id == element.id);
+                                      if (doctorSelected!.isActive) {
+                                        context.read<CitaBloc>().add(
+                                              CitaEvent.getCitas(
+                                                CitaRequestViewModel(
+                                                  doctorId: doctorSelected!.id,
+                                                  ubicacionesId: usuarioBloc.state.usuario?.ubicaciones ?? [],
+                                                  fechaHoraCita: DateTime.now(),
+                                                ),
+                                              ),
+                                            );
+                                      } else {
+                                        context.read<CitaBloc>().add(CitaEvent.invalidCita("Este doctor(a) está inactivo"));
+                                      }
+                                    }
+                                  }
+                                });
+                              }
+                            },
+                            child: DoctorCarousel(
+                              doctorIdInitialSelected: doctorSelected?.id,
+                              doctors: stt.doctors,
+                              onChanged: (doctor) {
+                                timerStart(timerCubit, () {
+                                  doctorSelected = doctor;
+                                  usuarioBloc.setDoctorSelected(doctor);
+                                  if (!doctor.isActive) {
+                                    context.read<CitaBloc>().add(CitaEvent.invalidCita("Este doctor(a) está inactivo"));
+                                    return;
+                                  }
+                                  context.read<CitaBloc>().add(CitaEvent.getCitas(
+                                        CitaRequestViewModel(
+                                          doctorId: doctor.id,
+                                          ubicacionesId: usuarioBloc.state.usuario?.ubicaciones ?? [],
+                                          fechaHoraCita: dateSelected,
+                                        ),
+                                      ));
+                                  context.read<CountIsolateCubit>().resetExternal();
+                                });
+                              },
+                            ),
                           );
                         },
                         failure: (stt) {
@@ -150,36 +204,41 @@ class CitaListDoctorPage extends StatelessWidget with ResponsiveWidgetMixin {
                                     fechaHoraCita: value,
                                   ),
                                 ));
+                                context.read<CountIsolateCubit>().resetExternal();
                               });
                             }
                           }
                         },
                       ),
                     ),
-                    // Positioned(
-                    //   left: 0,
-                    //   bottom: 0,
-                    //   top: 0,
-                    //   child: Stack(
-                    //     children: [
-                    //       const Icon(
-                    //         Icons.restart_alt,
-                    //         size: 32,
-                    //         color: AppColors.grey,
-                    //       ),
-                    //       Positioned.fill(
-                    //         child: Container(
-                    //           alignment: Alignment.center,
-                    //           child: AppTextGlobal.lightText(
-                    //             text: "$state",
-                    //             fontSize: 10,
-                    //             colorText: AppColors.grey,
-                    //           ),
-                    //         ),
-                    //       ),
-                    //     ],
-                    //   ),
-                    // )
+                    Positioned(
+                      left: 0,
+                      bottom: 0,
+                      top: 0,
+                      child: BlocBuilder<CountIsolateCubit, int>(
+                        builder: (context, state) {
+                          return Stack(
+                            children: [
+                              const Icon(
+                                Icons.restart_alt,
+                                size: 32,
+                                color: AppColors.grey,
+                              ),
+                              Positioned.fill(
+                                child: Container(
+                                  alignment: Alignment.center,
+                                  child: AppTextGlobal.lightText(
+                                    text: "$state",
+                                    fontSize: 10,
+                                    colorText: AppColors.grey,
+                                  ),
+                                ),
+                              ),
+                            ],
+                          );
+                        },
+                      ),
+                    )
                   ],
                 ),
               ),
@@ -221,7 +280,7 @@ class CitaListDoctorPage extends StatelessWidget with ResponsiveWidgetMixin {
               },
               citaLoaded: (state) {
                 // loaderCubit.hidden();
-                return CitasGroupedByHourDoctor(
+                return CitasGroupedByHourAsistRecep(
                   citas: state.citas,
                   onAdd: (hora, horaString) {
                     context.read<CitaCreateBloc>().add(CitaCreateEvent.citaPreCreateLocal(doctorSelected, dateSelected, hora, horaString));
