+++ lib/ui/modules/cita/r_asistente_recepcion/widgets/doctor_carrusel_card.dart
