+++ linux/flutter/generated_plugins.cmake
@@ -4,6 +4,7 @@
 
 list(APPEND FLUTTER_PLUGIN_LIST
   flutter_secure_storage
+  window_size
 )
 
 list(APPEND FLUTTER_FFI_PLUGIN_LIST
