+++ windows/flutter/generated_plugins.cmake
@@ -3,6 +3,7 @@
 #
 
 list(APPEND FLUTTER_PLUGIN_LIST
+  window_size
 )
 
 list(APPEND FLUTTER_FFI_PLUGIN_LIST
